+++ R/recomb_chr.R
@@ -1,5 +1,5 @@
-# parents are assumed to be data.frames with 2 columns: end, anc
-# (start is 0 for first chunk, end from previous row otherwise)
+# parents are assumed to be data.frames with 2 columns: posg, anc
+# (start is 0 for first chunk, posg from previous row otherwise)
 # mat/pat are symmetric (no distinction in practice)
 recomb_chr <- function( breaks, mat, pat ) {
     # validate data
@@ -19,20 +19,20 @@ recomb_chr <- function( breaks, mat, pat ) {
     # validate mat
     if ( !is.data.frame( mat ) )
         stop( '`mat` must be a data.frame (including tibble)' )
-    if ( !all( c('end', 'anc') %in% names( mat ) ) )
-        stop( '`mat` must have column names "end" and "anc"!' )
+    if ( !all( c('posg', 'anc') %in% names( mat ) ) )
+        stop( '`mat` must have column names "posg" and "anc"!' )
     if ( nrow( mat ) < 1 )
         stop( '`mat` must have at least one row!' )
     # validate pat
     if ( !is.data.frame( pat ) )
         stop( '`pat` must be a data.frame (including tibble)' )
-    if ( !all( c('end', 'anc') %in% names( pat ) ) )
-        stop( '`pat` must have column names "end" and "anc"!' )
+    if ( !all( c('posg', 'anc') %in% names( pat ) ) )
+        stop( '`pat` must have column names "posg" and "anc"!' )
     if ( nrow( pat ) < 1 )
         stop( '`pat` must have at least one row!' )
     # make sure chromosomes have the same length
-    leng <- mat$end[ nrow( mat ) ]
-    if ( pat$end[ nrow( pat ) ] != leng )
+    leng <- mat$posg[ nrow( mat ) ]
+    if ( pat$posg[ nrow( pat ) ] != leng )
         stop( '`mat` and `pat` chomosomes must be same length!' )
     # and make sure breaks are consistent with this length
     if ( any( breaks > leng ) )
@@ -73,18 +73,18 @@ recomb_chr <- function( breaks, mat, pat ) {
         }
         # identify rows to copy
         # bottom range first
-        while ( p$end[ i_start ] < break_last )
+        while ( p$posg[ i_start ] < break_last )
             i_start <- i_start + 1
         # top of range now
         i_end <- i_start
-        while ( p$end[ i_end ] < break_curr )
+        while ( p$posg[ i_end ] < break_curr )
             i_end <- i_end + 1
 
         # extract rows to pass to child for this chunk
         child_tmp <- p[ i_start:i_end, ]
         # since starts are implicit, if they get "shorter" there's no need here to edit that (they start where the previous chunk ends)
         # however, last end must be truncated to recombination position
-        child_tmp$end[ nrow( child_tmp ) ] <- break_curr
+        child_tmp$posg[ nrow( child_tmp ) ] <- break_curr
         # append to rest of child's chunks
         child <- rbind( child, child_tmp )