+++ R/recomb_init_founders.R
@@ -6,11 +6,17 @@
 #' @param ids The list of IDs to use for each individual
 #' @param lengs The lengths of each chromosome in centi-Morgans (cM).
 #' If this vector is named, the output inherits these chromosome names.
+#' If it is a list, it is assumed to be a recombination map (see [recomb_map_hg()] for examples) and the desired lengths extracted automatically (taken as the last value of column "posg" of each chromosome).
 #'
-#' @return A named list of diploid individuals, each of which is a list with two haploid individuals named "pat" and "mat", each of which is a list of chromosomes (inherits names of `lengs` if present), each of which is a tibble with a single row and two columns: "end" equals the chromosome length, and "anc" equals the ID of the individual (from `ids`) concatenated with either "_pat" or "_mat" depending on which parent it is.
+#' @return A named list of diploid individuals, each of which is a list with two haploid individuals named "pat" and "mat", each of which is a list of chromosomes (inherits names of `lengs` if present), each of which is a tibble with a single row and two columns: "posg" equals the chromosome length, and "anc" equals the ID of the individual (from `ids`) concatenated with either "_pat" or "_mat" depending on which parent it is.
 #'
 #' @examples
+#' # version with explicit recombination lengths
 #' ancs <- recomb_init_founders( c('a', 'b'), c(100, 200) )
+#' ancs
+#'
+#' # version using genetic map (uses provided human map) from which lengths are extracted
+#' ancs <- recomb_init_founders( c('a', 'b'), recomb_map_hg38 )
 #' ancs
 #'
 #' @seealso
@@ -25,6 +31,10 @@ recomb_init_founders <- function( ids, lengs ) {
         stop( '`lengs` is required!' )
     # `ids` can be almost anything, no need to test further
     # validate lengs
+    if ( is.list( lengs ) ) {
+        # extract from recombination map if that was provided
+        lengs <- recomb_map_lengs( lengs, name = 'lengs' )
+    }
     if ( !is.numeric( lengs ) )
         stop( '`lengs` must be numeric!' )
     if ( any( lengs <= 0 ) )
@@ -47,7 +57,7 @@ recomb_init_founders <- function( ids, lengs ) {
             for ( chr in 1 : n_chr ) {
                 # output tibble
                 hap[[ chr ]] <- tibble::tibble(
-                                               end = lengs[ chr ],
+                                               posg = lengs[ chr ],
                                                anc = id_par
                                            )
             }